+++ lib/src/animated_surface.dart
@@ -29,7 +29,7 @@ class AnimatedSurface extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final resolvedDuration = MediaQuery.of(context).disableAnimations
+    final resolvedDuration = MediaQuery.disableAnimationsOf(context)
         ? Duration.zero
         : duration;
     return AnimatedContainer(