+++ lib/src/slide_toggle.dart
@@ -37,7 +37,7 @@ class SlideToggle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final resolvedDuration = MediaQuery.of(context).disableAnimations
+    final resolvedDuration = MediaQuery.disableAnimationsOf(context)
         ? Duration.zero
         : duration;
     return AnimatedSlide(